Refactor: exception/handler 패키지 생성 [#39]
