+++ dashboard/src/main/java/whale/dashboard/exception/handler/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package whale.dashboard.exception;
+package whale.dashboard.exception.handler;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+import whale.dashboard.exception.EntityNotFoundException;
 
 import javax.validation.ConstraintViolationException;
